+++ src/lib/release.ts
@@ -1,7 +1,7 @@
 import packageInfo from "../../package.json";
 
 export const appVersion = packageInfo.version;
-export const releaseDate = "2026-09-08";
+export const releaseDate = "2026-09-20";
 
 export const releaseChannel = appVersion.includes("beta")
   ? "Beta 候选版"